[CLI] apply pre-commit formatting (black + EOF + line limit)
CI's pre-commit pass reformatted four files via black, trimmed the
trailing newline on test_docker_run_command.py, and required moving
the three new podman socket tests out of test_openclaw_shared_network.py
because adding them pushed that file past the project's 800-line
file-length limit. Split them into test_openclaw_podman_socket.py;
the helpers (_capture_run_commands, _find_container_run_cmd) are
imported from the original module.

No behavior change.
