+++ src/vllm-sr/tests/test_openclaw_shared_network.py
@@ -1,7 +1,5 @@
 from types import SimpleNamespace
 
-import os
-
 import pytest
 from cli import core, docker_cli, docker_services, docker_start, runtime_lifecycle
 from cli.runtime_stack import resolve_runtime_stack
@@ -263,148 +261,6 @@ def test_docker_start_vllm_sr_mounts_host_docker_cli_when_requested(
     dashboard_cmd = _find_container_run_cmd(captured, "vllm-sr-dashboard-container")
     assert "OPENCLAW_CONTAINER_RUNTIME=/usr/local/bin/docker" in dashboard_cmd
     assert f"{docker_bin}:/usr/local/bin/docker:ro" in dashboard_cmd
-
-
-def test_docker_start_vllm_sr_mounts_podman_socket_at_canonical_path(
-    tmp_path, monkeypatch
-):
-    """Podman runtime should mount podman.sock at /var/run/docker.sock and skip docker-CLI mount."""
-    config_path = tmp_path / "config.yaml"
-    config_path.write_text(
-        "version: v0.1\nlisteners:\n  - name: http-8899\n    address: 0.0.0.0\n    port: 8899\n"
-    )
-
-    podman_socket = tmp_path / "podman.sock"
-    podman_socket.write_text("")
-    monkeypatch.setenv("VLLM_SR_DOCKER_SOCKET", str(podman_socket))
-
-    monkeypatch.setattr(docker_start, "get_container_runtime", lambda: "podman")
-    monkeypatch.setattr(
-        docker_start,
-        "get_runtime_images",
-        lambda **kwargs: {
-            "router": "test-image",
-            "envoy": "test-image",
-            "dashboard": "test-image",
-        },
-    )
-    captured = _capture_run_commands(monkeypatch)
-
-    rc, _, _ = docker_cli.docker_start_vllm_sr(
-        str(config_path),
-        {},
-        [{"name": "http-8899", "address": "0.0.0.0", "port": 8899}],
-        network_name="vllm-sr-network",
-        openclaw_network_name="vllm-sr-network",
-        minimal=False,
-    )
-
-    assert rc == 0
-    dashboard_cmd = _find_container_run_cmd(captured, "vllm-sr-dashboard-container")
-    # podman.sock is mounted at the canonical /var/run/docker.sock so the in-image
-    # docker CLI in the dashboard transparently drives podman.
-    assert f"{podman_socket}:/var/run/docker.sock" in dashboard_cmd
-    # Dashboard backend keeps exec'ing "docker"; the in-image CLI is what runs.
-    assert "OPENCLAW_CONTAINER_RUNTIME=docker" in dashboard_cmd
-    # Host docker CLI should NOT be mounted in podman mode — the runtime is podman,
-    # not docker, so a host docker bin is irrelevant.
-    assert not any(
-        ":/usr/local/bin/docker:ro" in part for part in dashboard_cmd
-    ), f"unexpected host CLI mount in podman mode: {dashboard_cmd!r}"
-
-
-def test_docker_start_vllm_sr_resolves_default_podman_socket(tmp_path, monkeypatch):
-    """Without explicit override, podman mode should look up /run/podman/podman.sock."""
-    config_path = tmp_path / "config.yaml"
-    config_path.write_text(
-        "version: v0.1\nlisteners:\n  - name: http-8899\n    address: 0.0.0.0\n    port: 8899\n"
-    )
-
-    # Pretend /run/podman/podman.sock exists, /var/run/docker.sock does not.
-    monkeypatch.delenv("VLLM_SR_DOCKER_SOCKET", raising=False)
-    real_exists = os.path.exists
-
-    def fake_exists(path: str) -> bool:
-        if path == "/run/podman/podman.sock":
-            return True
-        if path in ("/var/run/docker.sock",):
-            return False
-        return real_exists(path)
-
-    monkeypatch.setattr(os.path, "exists", fake_exists)
-
-    monkeypatch.setattr(docker_start, "get_container_runtime", lambda: "podman")
-    monkeypatch.setattr(
-        docker_start,
-        "get_runtime_images",
-        lambda **kwargs: {
-            "router": "test-image",
-            "envoy": "test-image",
-            "dashboard": "test-image",
-        },
-    )
-    captured = _capture_run_commands(monkeypatch)
-
-    rc, _, _ = docker_cli.docker_start_vllm_sr(
-        str(config_path),
-        {},
-        [{"name": "http-8899", "address": "0.0.0.0", "port": 8899}],
-        network_name="vllm-sr-network",
-        openclaw_network_name="vllm-sr-network",
-        minimal=False,
-    )
-
-    assert rc == 0
-    dashboard_cmd = _find_container_run_cmd(captured, "vllm-sr-dashboard-container")
-    assert "/run/podman/podman.sock:/var/run/docker.sock" in dashboard_cmd
-
-
-def test_docker_start_vllm_sr_warns_when_podman_socket_missing(
-    tmp_path, monkeypatch, caplog
-):
-    """Podman mode should not fail if no podman socket is found — just warn."""
-    config_path = tmp_path / "config.yaml"
-    config_path.write_text(
-        "version: v0.1\nlisteners:\n  - name: http-8899\n    address: 0.0.0.0\n    port: 8899\n"
-    )
-
-    monkeypatch.delenv("VLLM_SR_DOCKER_SOCKET", raising=False)
-    real_exists = os.path.exists
-
-    def fake_exists(path: str) -> bool:
-        if "podman.sock" in path or "docker.sock" in path:
-            return False
-        return real_exists(path)
-
-    monkeypatch.setattr(os.path, "exists", fake_exists)
-
-    monkeypatch.setattr(docker_start, "get_container_runtime", lambda: "podman")
-    monkeypatch.setattr(
-        docker_start,
-        "get_runtime_images",
-        lambda **kwargs: {
-            "router": "test-image",
-            "envoy": "test-image",
-            "dashboard": "test-image",
-        },
-    )
-    captured = _capture_run_commands(monkeypatch)
-
-    rc, _, _ = docker_cli.docker_start_vllm_sr(
-        str(config_path),
-        {},
-        [{"name": "http-8899", "address": "0.0.0.0", "port": 8899}],
-        network_name="vllm-sr-network",
-        openclaw_network_name="vllm-sr-network",
-        minimal=False,
-    )
-
-    assert rc == 0
-    dashboard_cmd = _find_container_run_cmd(captured, "vllm-sr-dashboard-container")
-    # No socket mount when none exists.
-    assert not any("docker.sock" in part for part in dashboard_cmd), (
-        f"unexpected socket mount: {dashboard_cmd!r}"
-    )
 
 
 def test_docker_start_vllm_sr_mounts_dashboard_data_dir(tmp_path, monkeypatch):
